refactor helper.py to have meaningful class name
